@@ -8,14 +8,14 @@ if(project.hasProperty("matrix_malos.signing") && new File(project.property("mat
 }
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "admobilize.matrix.malosclient"
         minSdkVersion 19
-        targetSdkVersion 27
-        versionCode 399
-        versionName "1.2.1"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 28
+        versionCode 407
+        versionName "1.2.2"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -32,7 +32,7 @@ android {
             // dev utilizes minSDKVersion = 21 to allow the Android gradle plugin
             // to pre-dex each module and produce an APK that can be tested on
             // Android Lollipop without time consuming dex merging processes.
-            minSdkVersion 23
+            minSdkVersion 28
         }
         prod {
             minSdkVersion 19
@@ -54,8 +54,8 @@ android {
 dependencies {
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.google.android.material:material:1.0.0'
     implementation 'org.zeromq:jeromq:0.3.5'
     implementation 'io.grpc:grpc-okhttp:1.10.0'
     implementation 'io.grpc:grpc-protobuf-lite:1.10.0'
@@ -63,13 +63,13 @@ dependencies {
     implementation 'com.iamhabib:easy-preference:1.0.1'
     implementation 'javax.annotation:javax.annotation-api:1.2'
 
-    implementation ('com.crashlytics.sdk.android:crashlytics:2.6.8@aar') {
+    implementation ('com.crashlytics.sdk.android:crashlytics:2.7.1@aar') {
         transitive = true;
     }
 
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 }
 
 protobuf {
