@@ -1,8 +1,8 @@
 package admobilize.matrix.malosclient.ui;
 
 import android.os.Bundle;
-import android.os.HandlerThread;
-import android.support.v4.app.DialogFragment;
+
+import androidx.fragment.app.DialogFragment;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
